Moved setTimeout to specific functions and removed setInterval
Rather than putting setTimeout calls around function calls and the chance of not wrapping a call, I've moved the logic into the function themselves so it can be removed in the future and there is no need for the caller to take that into account.

Removed setInterval as that would add an action to the event loop even if the browser was busy. Used setTimeout and calling itself so that it will only add itself to the event loop if it has just finished executing.